Fixed areas in code where double && was missing and just a & or where…

+++ src/pawgui/pawgui_base.cpp
@@ -85,7 +85,7 @@ namespace pawgui
 
         int foundPosition = lowered_string.find( s_name );
 
-        if ( foundPosition!=std::string::npos )
+        if ( foundPosition!=(int)std::string::npos )
         {
             return true;
         }
